+++ grpc_extra/model/__init__.py
@@ -1,12 +1,21 @@
 from .data_helper import DefaultModelDataHelper, ModelDataHelper
 from .filtering import ModelFilterSchema
-from .schemas import AllowedEndpoints, ModelServiceConfig
+from .schemas import (
+    AllowedEndpoints,
+    ChoiceEndpointConfig,
+    IntChoiceSchema,
+    ModelServiceConfig,
+    TextChoiceSchema,
+)
 from .service import ModelService
 
 __all__ = [
     "ModelService",
     "ModelServiceConfig",
     "AllowedEndpoints",
+    "ChoiceEndpointConfig",
+    "IntChoiceSchema",
+    "TextChoiceSchema",
     "ModelDataHelper",
     "DefaultModelDataHelper",
     "ModelFilterSchema",